decomp: fix 3 MakeString template instantiations to 100%
HamCamShot: pass Symbol directly instead of .Str() in MakeString.
DataEventList: cast to (char*) for correct template deduction.
SampleData: change max rev constants from int to unsigned short.

+++ src/system/synth/SampleData.cpp
@@ -11,8 +11,8 @@
 
 SampleDataAllocFunc SampleData::sAlloc = nullptr;
 SampleDataFreeFunc SampleData::sFree = nullptr;
-static const int gSampleDataMaxRev = 0x10;
-static const int gSampleDataMaxAltRev = 0;
+static const unsigned short gSampleDataMaxRev = 0x10;
+static const unsigned short gSampleDataMaxAltRev = 0;
 
 SampleData::SampleData() : mData(0), mMarkers() { Reset(); }
 SampleData::~SampleData() { Dealloc(); }
